move ApolloRequest to the apollo package
