+++ apollo-runtime-kotlin/src/commonMain/kotlin/com/apollographql/apollo/network/NetworkTransport.kt
@@ -4,7 +4,7 @@ import com.apollographql.apollo.api.ApolloExperimental
 import com.apollographql.apollo.api.CustomScalarAdapters
 import com.apollographql.apollo.api.ExecutionContext
 import com.apollographql.apollo.api.Operation
-import com.apollographql.apollo.interceptor.ApolloRequest
+import com.apollographql.apollo.ApolloRequest
 import com.apollographql.apollo.interceptor.ApolloResponse
 import kotlinx.coroutines.flow.Flow
