+++ apollo-runtime-kotlin/src/commonTest/kotlin/com/apollographql/apollo/OauthInterceptorTest.kt
@@ -6,7 +6,6 @@ import com.apollographql.apollo.api.ExecutionContext
 import com.apollographql.apollo.api.Operation
 import com.apollographql.apollo.api.parse
 import com.apollographql.apollo.exception.ApolloHttpException
-import com.apollographql.apollo.interceptor.ApolloRequest
 import com.apollographql.apollo.interceptor.ApolloResponse
 import com.apollographql.apollo.interceptor.BearerTokenInterceptor
 import com.apollographql.apollo.network.HttpExecutionContext
