+++ apollo-testing-support/src/commonMain/kotlin/com/apollographql/apollo/testing/MockNetworkTransport.kt
@@ -5,7 +5,7 @@ import com.apollographql.apollo.api.CustomScalarAdapters
 import com.apollographql.apollo.api.ExecutionContext
 import com.apollographql.apollo.api.Operation
 import com.apollographql.apollo.api.parse
-import com.apollographql.apollo.interceptor.ApolloRequest
+import com.apollographql.apollo.ApolloRequest
 import com.apollographql.apollo.interceptor.ApolloResponse
 import com.apollographql.apollo.network.NetworkTransport
 import kotlinx.coroutines.ExperimentalCoroutinesApi
