+++ apollo-testing-support/src/commonMain/kotlin/com/apollographql/apollo/testing/TestLoggerExecutor.kt
@@ -3,7 +3,7 @@ package com.apollographql.apollo.testing
 import com.apollographql.apollo.api.ApolloExperimental
 import com.apollographql.apollo.api.Operation
 import com.apollographql.apollo.interceptor.ApolloInterceptorChain
-import com.apollographql.apollo.interceptor.ApolloRequest
+import com.apollographql.apollo.ApolloRequest
 import com.apollographql.apollo.interceptor.ApolloRequestInterceptor
 import com.apollographql.apollo.interceptor.ApolloResponse
 import kotlinx.coroutines.ExperimentalCoroutinesApi
